+++ Src/Noesis/Core/Src/Proxies/Shape.cs
@@ -158,9 +158,9 @@ public class Shape : FrameworkElement {
     } 
   }
 
-  public float StrokeDashOffset {
+  public double StrokeDashOffset {
     set {
-      NoesisGUI_PINVOKE.Shape_StrokeDashOffset_set(swigCPtr, value);
+      NoesisGUI_PINVOKE.Shape_StrokeDashOffset_set(swigCPtr, (float)value);
     } 
     get {
       float ret = NoesisGUI_PINVOKE.Shape_StrokeDashOffset_get(swigCPtr);
@@ -188,9 +188,9 @@ public class Shape : FrameworkElement {
     } 
   }
 
-  public float StrokeMiterLimit {
+  public double StrokeMiterLimit {
     set {
-      NoesisGUI_PINVOKE.Shape_StrokeMiterLimit_set(swigCPtr, value);
+      NoesisGUI_PINVOKE.Shape_StrokeMiterLimit_set(swigCPtr, (float)value);
     } 
     get {
       float ret = NoesisGUI_PINVOKE.Shape_StrokeMiterLimit_get(swigCPtr);
@@ -208,9 +208,9 @@ public class Shape : FrameworkElement {
     } 
   }
 
-  public float StrokeThickness {
+  public double StrokeThickness {
     set {
-      NoesisGUI_PINVOKE.Shape_StrokeThickness_set(swigCPtr, value);
+      NoesisGUI_PINVOKE.Shape_StrokeThickness_set(swigCPtr, (float)value);
     } 
     get {
       float ret = NoesisGUI_PINVOKE.Shape_StrokeThickness_get(swigCPtr);
